Create StatementOfMeansFeature#isApplicableFor to check whether SoM is available in task builder, SoM task and guards
