+++ src/main/features/response/routes/full-admission/payment-option.ts
@@ -10,9 +10,9 @@ import { User } from 'main/app/idam/user'
 import { DraftService } from 'services/draftService'
 import { FeatureToggleGuard } from 'main/app/guards/featureToggleGuard'
 import { ResponseDraft } from 'response/draft/responseDraft'
-import { StatementOfMeans } from 'response/draft/statementOfMeans'
 import { Draft } from '@hmcts/draft-store-client'
 import { Claim } from 'main/app/claims/models/claim'
+import { StatementOfMeansFeature } from 'response/helpers/statementOfMeansFeature'
 
 function renderView (form: Form<DefendantPaymentOption>, res: express.Response) {
   const draft: Draft<ResponseDraft> = res.locals.responseDraft
@@ -21,7 +21,7 @@ function renderView (form: Form<DefendantPaymentOption>, res: express.Response) 
     form: form,
     claim: claim,
     draft: draft.document,
-    statementOfMeansIsApplicable: StatementOfMeans.isApplicableFor(draft.document)
+    statementOfMeansIsApplicable: StatementOfMeansFeature.isApplicableFor(draft.document)
   })
 }
